Move tests outside of package; keep only tests related to bot functionality, remove API tests
